pyproject.toml
[project]
name = "e-sign-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "docx2pdf>=0.1.8",
    "fastapi>=0.135.3",
    "jinja2>=3.1.6",
    "mcp>=1.27.0",
    "pillow>=12.2.0",
    "pydantic>=2.13.0",
    "pydantic-settings>=2.13.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.26",
    "reportlab>=4.4.10",
    "uvicorn>=0.44.0",
]
